requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
defusedxml==0.7.1
dj-database-url==2.1.0
Django==5.1
django-bootstrap-icons==0.8.6
django-widget-tweaks==1.5.0
idna==3.8
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
slug==2.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
gdacs-api==2.0.0
fontawesomefree==6.6.0
